Fix for topology drag performance
