+++ frontend/packages/dev-console/src/components/topology/Topology.scss
@@ -1,0 +1,7 @@
+.odc-topology-graph-view {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  right: 0;
+  left: 0;
+}
