+++ frontend/packages/dev-console/src/components/topology/__tests__/TopologySideBar.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import { CloseButton } from '@console/internal/components/utils';
-import SideBar, { TopologySideBarProps } from '../TopologySideBar';
+import { TopologySideBar, TopologySideBarProps } from '../TopologySideBar';
 
 describe('TopologySideBar:', () => {
   const props: TopologySideBarProps = {
@@ -10,13 +10,13 @@ describe('TopologySideBar:', () => {
   };
 
   it('renders a SideBar', () => {
-    const wrapper = shallow(<SideBar {...props} />);
+    const wrapper = shallow(<TopologySideBar {...props} />);
     expect(wrapper.exists()).toBeTruthy();
   });
 
   it('clicking on close button should call the onClose callback function', () => {
     const onClose = jest.fn();
-    const wrapper = shallow(<SideBar show onClose={onClose} />);
+    const wrapper = shallow(<TopologySideBar show onClose={onClose} />);
     wrapper
       .find(CloseButton)
       .shallow()
